+++ vrp-pragmatic/src/constraints/groups.rs
@@ -38,37 +38,20 @@ impl ConstraintModule for GroupModule {
     fn accept_insertion(&self, solution_ctx: &mut SolutionContext, route_index: usize, job: &Job) {
         if let Some(group) = get_group(job) {
             let route_ctx = solution_ctx.routes.get_mut(route_index).unwrap();
-            let jobs_count = route_ctx.route.tour.job_count();
+
             let mut groups = get_groups(route_ctx);
             groups.insert(group.clone());
-            route_ctx.state_mut().put_route_state(self.state_key, (groups, jobs_count))
+
+            route_ctx.state_mut().put_route_state(self.state_key, groups)
         }
     }
 
-    fn accept_route_state(&self, ctx: &mut RouteContext) {
-        let current_jobs_count = ctx.route.tour.job_count();
-        let old_jobs_count = ctx
-            .state
-            .get_route_state::<(HashSet<String>, usize)>(self.state_key)
-            .map(|(_, jobs)| *jobs)
-            .unwrap_or(current_jobs_count);
-
-        if old_jobs_count != current_jobs_count {
-            let groups = get_groups(ctx);
-            ctx.state_mut().put_route_state(self.state_key, (groups, current_jobs_count))
-        }
-    }
+    fn accept_route_state(&self, _: &mut RouteContext) {}
 
     fn accept_solution_state(&self, solution_ctx: &mut SolutionContext) {
-        // NOTE we can filter here by stale flag, but then we need to keep non-changed route
-        // cache and identify routes which where deleted to remove them from cache. Instead,
-        // let's go through all routes and create evaluation cache from scratch. However, this
-        // approach has performance implications for calling `accept_solution_state` method frequently.
-
-        solution_ctx.routes.iter_mut().filter(|route_ctx| route_ctx.is_stale()).for_each(|route_ctx| {
-            let current_jobs_count = route_ctx.route.tour.job_count();
+        solution_ctx.routes.iter_mut().for_each(|route_ctx| {
             let groups = get_groups(route_ctx);
-            route_ctx.state_mut().put_route_state(self.state_key, (groups, current_jobs_count));
+            route_ctx.state_mut().put_route_state(self.state_key, groups);
         });
     }
 
@@ -103,30 +86,22 @@ impl HardRouteConstraint for GroupHardRouteConstraint {
         job: &Job,
     ) -> Option<RouteConstraintViolation> {
         get_group(job).and_then(|group| {
+            let is_partial_problem = solution_ctx.get_jobs_amount() != self.total_jobs;
+            if is_partial_problem {
+                return Some(RouteConstraintViolation { code: self.code });
+            }
+
             let other_route = solution_ctx
                 .routes
                 .iter()
                 .filter(|rc| rc.route.actor != route_ctx.route.actor)
-                .filter_map(|rc| rc.state.get_route_state::<(HashSet<String>, usize)>(self.state_key))
-                .any(|(groups, _)| groups.contains(group));
+                .filter_map(|rc| rc.state.get_route_state::<HashSet<String>>(self.state_key))
+                .any(|groups| groups.contains(group));
 
-            let current_route = route_ctx
-                .state
-                .get_route_state::<(HashSet<String>, usize)>(self.state_key)
-                .map_or(false, |(groups, _)| groups.contains(group));
-
-            match (other_route, current_route) {
-                (true, _) => Some(RouteConstraintViolation { code: self.code }),
-                // NOTE handle partial solution context use case (e.g. decompose search)
-                (false, false) => {
-                    let is_full_problem = solution_ctx.get_jobs_amount() == self.total_jobs;
-                    if is_full_problem {
-                        None
-                    } else {
-                        Some(RouteConstraintViolation { code: self.code })
-                    }
-                }
-                (false, true) => None,
+            if other_route {
+                Some(RouteConstraintViolation { code: self.code })
+            } else {
+                None
             }
         })
     }